pyrs: Avoid usize
Fixes https://github.com/py2many/py2many/issues/491

+++ tests/expected/bubble_sort.rs
@@ -27,7 +27,7 @@ use anyhow::Result;
 use std::collections;
 
 pub fn bubble_sort(seq: &mut Vec<i32>) -> Vec<i32> {
-    let L = seq.len();
+    let L = seq.len() as i32;
     for _ in (0..L) {
         for n in (1..L) {
             if seq[n as usize] < seq[((n as i32) - 1) as usize] {
